Redo db access almost completely.  Simplify panel switching.
Artists and Concerts now have a wrapper method which provides access
to fields from the database on-the-fly, which means updates to
fields on the screen (like browsedate) will be seen immediately.
This also may require slightly less memory, as the data (especially
names) are no longer kept in memory at all times.

A base class handles the database access, and is used to define
properties for the wrapper class. This provides a much simpler interface.

Also moved the responsibility of knowing which UI column corresponds to
which field up to the UI layer where it belongs.

In addition, changed the switching of panels in the main frame to
no longer need to know what the currently active frame is.  The new
frame simply replaces whatever is there.