Big reorg of how database is accessed. Step 1.
Moving away from having the database be a global asset. Make a bunch of
classes (ArtistList & Artist, ConcertList & Concert & ConcertDetails)
take a database handle as a constructor argument, which is passed
between them as necessary.

Since this increases the size of Artist and Concert objects, the ...List
classes now store just the db ids, internally, and construct the
appropriate object on the fly when necessary.

This is only a first pass, but the code seems to run, more-or-less, so
that's promising! :)