access lastupdate date and new count
AristList and ConcertList both now provide access to the lastupdate date
and the new records count. At the moment, this information is merely
shown on the main panels, but could have more uses in the future, such
as greying out the "Clear New" menu choice when the new count is zero.